+++ CHANGELOG.md
@@ -21,6 +21,13 @@
   (default) or `tar` format. Images pushed with `--layer-format tar` can be
   pulled and run by other OCI runtimes.
 
+### Bug Fixes
+
+- Fix fall-back to temporary sandbox rootfs bundle in OCI-Mode for OCI URIs
+  (`docker://`) etc.
+- Fix confusing error messages / incorrect fall-back attempt when explicit
+  execution of an OCI-SIF fails.
+
 ### Requirements
 
 - Requires a minimum of Go 1.21.5 to build due to dependency updates.